functions/requirements.txt
firebase-admin==6.4.0
firebase-functions==0.4.0
google-cloud-speech==2.26.0
